New LED system + Retuning shooters + New Middle Lift
Installed LED subsystems to convey what color ball the color sensors are reading in each chute. Retuned flyShooters to put backspin on the balls and dampened the power. Also put offsets in the lifts to accommodate for different launch speeds of the balls.  Also added some logic to RedAuto to take obelisk April Tag and turn it into launch sequences for identification points.